Settings: On-the-spot preference clear option
Allow for handling of resetting multiple preferences individually.

Present the user with a list of active on-the-spot preferences to
choose from.
If a preference dialog was never presented to the user, it will not
be available for reset.
If only one preference is currently active, a simple reset dialog is
presented (current existing dialog / no list)

PS1: Initial commit
PS2: Use defined strings instead of decoding the SETTINGS_TO_RESET variables.
     This allows for string translation.
     String names should be equivalent to the SETTINGS_TO_RESET variable.
PS3: Add some error checks

Change-Id: If9c483e0745000ab486f060bf87756204b4a8e6e

Conflicts:
	src/com/android/settings/PrivacySettings.java